sprawadza czy plik juz istnieje i typ pliku

+++ pub/sandbox/index.php
@@ -19,10 +19,27 @@
     if(isset($_POST['submit']))
     {
         $targetDir = "img/";
+
         $sourceFileName = $_FILES['uploadedFile']['name'];
+
         $tempURL = $_FILES['uploadedFile']['tmp_name'];
+
+        $imgInfo = getimagesize($tempURL);
+        if(!is_array($imgInfo)) 
+        {
+            die("BŁĄD: Przekazany plik nie jest obrazem!");
+        }
+
         $targetURL = $targetDir . $sourceFileName;
+
+        if(file_exists($targetURL)) 
+        {
+            die("BŁĄD: Podany plik już istnieje!");
+        }
+
         move_uploaded_file($tempURL, $targetURL);
+
+        echo "Plik został poprawnie wgrany na serwer";
     }
     ?>
 </body>
